refactor: make providers directory
