+++ client/src/App.tsx
@@ -3,8 +3,8 @@ import HelloWeb3 from "./components/pages/HelloWeb3";
 import SignUp from "./components/pages/SignUp";
 import SignIn from "./components/pages/SignIn";
 import Mypage from "./components/pages/Mypage";
-import OptionalAuthProvider from "./components/organisms/OptionalAuthProvider";
-import AuthProvider from "./components/organisms/AuthProvider";
+import OptionalAuthProvider from "./components/organisms/providers/OptionalAuthProvider";
+import AuthProvider from "./components/organisms/providers/AuthProvider";
 import ImageUpload from "./components/pages/ImageUpload";
 import Header from "./components/organisms/Header";
 import NotFound from "./components/pages/NofFound";
