+++ client/src/components/pages/ImageUpload/index.tsx
@@ -1,7 +1,7 @@
 import { PutObjectCommand } from "@aws-sdk/client-s3";
 import { fromCognitoIdentityPool } from "@aws-sdk/credential-providers";
 import { useCallback, useMemo, useState } from "react";
-import { useAuth } from "~/components/organisms/AuthProvider";
+import { useAuth } from "~/components/organisms/providers/AuthProvider";
 import { COGNITO_USER_POOL_ID } from "~/configs/cognito";
 import { COGNITO_IDENTITY_POOL_ID, getS3Client } from "~/configs/s3";
